Core Pigulator Design
This is a basic math concept for the Pigulator. It is in a very early setup with no UI and has no way to validate user input. This means if you put in the wrong values the system will throw an exception and close, or calculate unattainable numbers. It is currently possible to have customers pay more than 100% if you input the wrong numbers for percentage pay.